refactor route /start to / for simplicity
