+++ app/(ui)/signedin/page.tsx
@@ -37,12 +37,12 @@ export default async function Page(props: { searchParams: Promise<SearchParams> 
 
     // Verify the session is fully authenticated
     if (!sessionFactors) {
-      redirect("/start");
+      redirect("/");
     }
 
     const valid = await isSessionValid({ serviceUrl, session: sessionFactors });
     if (!valid) {
-      redirect("/start");
+      redirect("/");
     }
 
     if (!requestId) {
@@ -56,7 +56,7 @@ export default async function Page(props: { searchParams: Promise<SearchParams> 
     if (isRedirectError(error)) {
       throw error;
     }
-    redirect("/start");
+    redirect("/");
   }
 
   return (
